requirements.txt
streamlit
selenium
beautifulsoup4
pandas
requests-html
validators
webdriver-manager
openai
python-dotenv